size_t/ptrdiff_t transition part 2: SFIO, hash lib, fmt*(), strmatch() (#941)
This is the second of the thickfold patch series, which enables
ksh93 to operate within a 64-bit address space. This part of the
patch series includes everything affected by changes to <ast.h>.

The parts of ksh93 affected by this commit are:
- SFIO and the associated stdio wrapper. The sfvprintf and sfvscanf
  functions are notable for bearing the brunt of the substantial
  code changes.
  - Nearly all of the code is de novo, though towards the latter
    end of development I used graphviz commits for cross-reference:
    - https://gitlab.com/graphviz/graphviz/-/commit/6451a669
    - https://gitlab.com/graphviz/graphviz/-/commit/153a8f87
    - Improved multibyte handling in sfvscanf via use of
      unsigned char*; ported from graphviz:
      https://gitlab.com/graphviz/graphviz/-/commit/cb9b35d1
      (I'm aware this function is unused, but after some manual
      testing it works well AFAICT. A bit odd that ksh93 never
      uses it once.)
  - sfseek(): Removed a wasteful double assignment for f->iosz.
    This error was introduced in 2003-06-21 ksh93o+, and likely
    would have gone unnoticed if not for the thickfold project.
    (In practice this line was virtually always optimized out by
    the compiler, so it was dead code.)

- 64-bit modernization for the AST hash library.
- Transitioned pathgetlink() to return ssize_t.
- Updated the code in print.c for compatibility with the SFIO
  64-bit modernization.
- Transitioned the strgrpmatch() family to use regflags_t to
  appease some -Wsign-conversion warnings (the type is now located
  in ast.h).
  - The now unused STR_INT flag backported from ksh93v- has been
    removed to avoid bitrot.
- Removed the unused nonstandard ssizeof() macro from ast.h.
- Made the snprintf and vsnprintf wrappers standards compliant with
  the C standard.

Change in the number of warnings on Linux when compiling with clang
using -Wsign-compare -Wshorten-64-to-32 -Wsign-conversion
-Wimplicit-int-conversion: 3,849 => 3,178 => 61 (progression from
part 1 => part 2 => part 13)

Progresses https://github.com/ksh93/ksh/issues/592

+++ src/lib/libast/hash/hashdump.c
@@ -2,7 +2,7 @@
 *                                                                      *
 *               This software is part of the ast package               *
 *          Copyright (c) 1985-2011 AT&T Intellectual Property          *
-*          Copyright (c) 2020-2023 Contributors to ksh 93u+m           *
+*          Copyright (c) 2020-2026 Contributors to ksh 93u+m           *
 *                      and is licensed under the                       *
 *                 Eclipse Public License, Version 2.0                  *
 *                                                                      *
@@ -14,6 +14,7 @@
 *                  David Korn <dgk@research.att.com>                   *
 *                   Phong Vo <kpv@research.att.com>                    *
 *                  Martijn Dekker <martijn@inlv.org>                   *
+*            Johnothan King <johnothanking@protonmail.com>             *
 *                                                                      *
 ***********************************************************************/
 /*
@@ -51,7 +52,7 @@ dumpbucket(Hash_table_t* tab, int flags)
 	Hash_bucket_t**		sp;
 	Hash_bucket_t*		b;
 	Hash_bucket_t**		sx;
-	int			n;
+	size_t			n;
 	unsigned char*		s;
 
 	NoP(flags);
@@ -64,7 +65,7 @@ dumpbucket(Hash_table_t* tab, int flags)
 				n++;
 		if (n)
 		{
-			sfprintf(sfstderr, "%5d %2d :", sp - tab->table, n);
+			sfprintf(sfstderr, "%5td %2zu :", sp - tab->table, n);
 			for (b = *sp; b; b = b->next)
 				if (!(b->hash & HASH_DELETED) && (!(tab->flags & HASH_VALUE) || b->value))
 				{
@@ -112,12 +113,12 @@ dumptable(Hash_table_t* tab, int flags)
 	sfprintf(sfstderr, "\n");
 	sfprintf(sfstderr, "        address:     0x%08lx\n", (unsigned long)tab);
 	sfprintf(sfstderr, "        flags:       ");
-	if (tab->frozen) sfprintf(sfstderr, "frozen=%d ", tab->frozen);
+	if (tab->frozen) sfprintf(sfstderr, "frozen=%u ", (unsigned int)tab->frozen);
 	dumpflags(tab->flags);
 	sfprintf(sfstderr, "\n");
-	sfprintf(sfstderr, "        size:        %d\n", tab->size);
+	sfprintf(sfstderr, "        size:        %jd\n", (intmax_t)tab->size);
 	sfprintf(sfstderr, "        buckets:     %d\n", tab->buckets);
-	sfprintf(sfstderr, "        bucketsize:  %d\n", tab->bucketsize * sizeof(char*));
+	sfprintf(sfstderr, "        bucketsize:  %zu\n", (size_t)tab->bucketsize * sizeof(char*));
 	sfprintf(sfstderr, "\n");
 	if ((flags | tab->flags) & HASH_BUCKET) dumpbucket(tab, flags);
 }
@@ -135,7 +136,7 @@ dumproot(Hash_root_t* root, int flags)
 	sfprintf(sfstderr, "        address:     0x%08lx\n", (unsigned long)root);
 	sfprintf(sfstderr, "        flags:       ");
 	dumpflags(root->flags);
-	if (root->namesize) sfprintf(sfstderr, "namesize=%d ", root->namesize);
+	if (root->namesize) sfprintf(sfstderr, "namesize=%zu ", root->namesize);
 	if (root->local->alloc) sfprintf(sfstderr, "alloc=0x%08lx ", (unsigned long)root->local->alloc);
 	if (root->local->compare) sfprintf(sfstderr, "compare=0x%08lx ", (unsigned long)root->local->compare);
 	if (root->local->free) sfprintf(sfstderr, "free=0x%08lx ", (unsigned long)root->local->free);
