+++ src/lib/libast/sfio/sfexcept.c
@@ -2,7 +2,7 @@
 *                                                                      *
 *               This software is part of the ast package               *
 *          Copyright (c) 1985-2011 AT&T Intellectual Property          *
-*          Copyright (c) 2020-2024 Contributors to ksh 93u+m           *
+*          Copyright (c) 2020-2026 Contributors to ksh 93u+m           *
 *                      and is licensed under the                       *
 *                 Eclipse Public License, Version 2.0                  *
 *                                                                      *
@@ -14,6 +14,7 @@
 *                  David Korn <dgk@research.att.com>                   *
 *                   Phong Vo <kpv@research.att.com>                    *
 *                  Martijn Dekker <martijn@inlv.org>                   *
+*            Johnothan King <johnothanking@protonmail.com>             *
 *                                                                      *
 ***********************************************************************/
 #include	"sfhdr.h"
@@ -75,7 +76,7 @@ int _sfexcept(Sfio_t*	f,	/* stream where the exception happened */
 			if((io -= size) <= 0)
 				io = SFIO_GRAIN;
 			size = ((size+io+SFIO_GRAIN-1)/SFIO_GRAIN)*SFIO_GRAIN;
-			if(!(data = realloc(f->data,size)))
+			if(!(data = realloc(f->data,(size_t)size)))
 				goto chk_stack;
 			f->endb = data + size;
 			f->next = data + (f->next - f->data);
